ModularScatter refactor ready for testing + csproj fixes csproj fixes : - added a separate MFI reference for the KSP 1.8 build (MFI 1.2.7) ModularScatter : - Should now be fully functional - HeatEmitter : removed some features that were added ~3 years ago, that have nothing to do with scatters, that doesn't work correctly, that were breaking the original functionality and that are apparently not used by any planet pack according to a github search. - Removed the "densityVariance" option : it wasn't implemented at all and its units were undocumented, so any attempt at "fixing" it would result in random scatter distribution changes in the many existing planets packs defining it.
