+++ dependencies/MFI_1.2.7_KSP1.8/ModularFlightIntegrator.version
@@ -1,0 +1,26 @@
+{
+    "NAME": "ModularFlightIntegrator",
+    "URL": "https://ksp.sarbian.com/jenkins/job/ModularFlightIntegrator/lastSuccessfulBuild/artifact/ModularFlightIntegrator/ModularFlightIntegrator.version",
+    "DOWNLOAD": "https://forum.kerbalspaceprogram.com/index.php?/topic/106369-x",
+    "VERSION":{
+        "MAJOR": 1,
+        "MINOR": 2,
+        "PATCH": 7,
+        "BUILD": 0
+    },
+    "KSP_VERSION": {
+        "MAJOR": 1,
+        "MINOR": 8,
+        "PATCH": 0
+    },
+    "KSP_VERSION_MIN": {
+        "MAJOR": 1,
+        "MINOR": 8,
+        "PATCH": 0
+    },
+    "KSP_VERSION_MAX": {
+        "MAJOR": 1,
+        "MINOR": 10,
+        "PATCH": 90
+    }
+}
