+++ src/Kopernicus/Configuration/ModularScatterLoader/HeatEmitter.cs
@@ -47,53 +47,8 @@ namespace Kopernicus.Configuration.ModularScatterLoader
         [KittopiaDescription("The ambient temperature.")]
         public NumericParser<Double> AmbientTemp
         {
-            get { return Value.ambientTemp; }
-            set { Value.ambientTemp = value; }
-        }
-
-        // If the ambientTemp should be added.
-        [ParserTarget("sumTemp")]
-        [KittopiaDescription("If the ambientTemp should be added.")]
-        public NumericParser<Boolean> SumTemp
-        {
-            get { return Value.sumTemp; }
-            set { Value.sumTemp = value; }
-        }
-
-        // The name of the biome.
-        [ParserTarget("biomeName")]
-        [KittopiaDescription("The name of the biome.")]
-        public String BiomeName
-        {
-            get { return Value.biomeName; }
-            set { Value.biomeName = value; }
-        }
-
-        // Multiplier curve to change ambientTemp with altitude
-        [ParserTargetCollection("AltitudeCurve", Key = "key", NameSignificance = NameSignificance.Key)]
-        [KittopiaDescription("Multiplier curve to change ambientTemp with altitude.")]
-        public List<NumericCollectionParser<Single>> AltitudeCurve
-        {
-            get { return Utility.FloatCurveToList(Value.altitudeCurve); }
-            set { Value.altitudeCurve = Utility.ListToFloatCurve(value); }
-        }
-
-        // Multiplier curve to change ambientTemp with latitude
-        [ParserTargetCollection("LatitudeCurve", Key = "key", NameSignificance = NameSignificance.Key)]
-        [KittopiaDescription("Multiplier curve to change ambientTemp with latitude.")]
-        public List<NumericCollectionParser<Single>> LatitudeCurve
-        {
-            get { return Utility.FloatCurveToList(Value.latitudeCurve); }
-            set { Value.latitudeCurve = Utility.ListToFloatCurve(value); }
-        }
-
-        // Multiplier curve to change ambientTemp with longitude
-        [ParserTargetCollection("LongitudeCurve", Key = "key", NameSignificance = NameSignificance.Key)]
-        [KittopiaDescription("Multiplier curve to change ambientTemp with longitude.")]
-        public List<NumericCollectionParser<Single>> LongitudeCurve
-        {
-            get { return Utility.FloatCurveToList(Value.longitudeCurve); }
-            set { Value.longitudeCurve = Utility.ListToFloatCurve(value); }
+            get { return Value.temperature; }
+            set { Value.temperature = value; }
         }
 
         // Multiplier curve to change ambientTemp with distance
@@ -103,15 +58,6 @@ namespace Kopernicus.Configuration.ModularScatterLoader
         {
             get { return Utility.FloatCurveToList(Value.distanceCurve); }
             set { Value.distanceCurve = Utility.ListToFloatCurve(value); }
-        }
-
-        // Multiplier map for ambientTemp
-        [ParserTarget("HeatMap")]
-        [KittopiaDescription("Greyscale map for fine control of the ambientTemp on a planet. black = 0, white = 1")]
-        public MapSOParserGreyScale<MapSO> HeatMap
-        {
-            get { return Value.heatMap; }
-            set { Value.heatMap = value; }
         }
     }
 }
